fix cascade confidence check and dtype regression
Three bugs killed performance and cascade hit rate:

1. torch_dtype deprecation (model.py, draft.py): newer transformers ignores
   the `torch_dtype` kwarg and silently loads fp32, causing 3-4x slowdown on
   Llama-3B. Fixed by using the current `dtype=` kwarg in both files.

2. Wrong probability for cascade gate (worker.py, draft.py): cascade checked
   `sampled_prob` (probability of the randomly-drawn token) instead of
   `max_prob` (peak of the softmax distribution). At temperature=1.0 a model
   that is 80% confident on one token can still sample a 15% token — the old
   check would see 0.15 and not fire even when the model is highly certain.
   Added `DraftModel.draft_peek()` which returns both values explicitly, and
   the cascade gate now checks `d0_prob_max >= threshold`.

3. cascade_hits counter not initialized in pre-cascade code path (minor):
   variable was read at generation end but only conditionally written.
   Guaranteed zero-init at loop start fixes the potential NameError.

Together these restore ~2.4 TPS on MPS (was 0.65 with dtype bug) and make
cascade actually fire at threshold=0.6.

+++ model.py
@@ -267,7 +267,7 @@ def _get_stage_llama(stage_id: int, num_stages: int, config: dict) -> nn.Module:
     model_name = config["model"]["name"]
     dtype = resolve_dtype(config)
     print(f"  Loading {model_name} ({dtype}) — stage {stage_id}/{num_stages - 1}...")
-    full = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=dtype)
+    full = AutoModelForCausalLM.from_pretrained(model_name, dtype=dtype)
     full.eval()
 
     num_layers = len(full.model.layers)
